requirements.txt
langchain>=1.0.0
langchain-community
langsmith>=0.4.37
langserve[all]
fastapi
uvicorn
python-dotenv
langchain-ollama 
pydantic
chromadb
ddgs  
langchain-ollama  
langgraph-cli   
ollama  
PyYAML 
requests
uvicorn